+++ seaborn_timestamp/time_profile.py
@@ -42,8 +42,8 @@ from time import sleep
 import logging
 
 log = logging.getLogger(__name__)
-from seaborn.meta.calling_function import function_info
-from seaborn.table.table import SeabornTable
+from seaborn_meta.calling_function import function_info
+from seaborn_table.table import SeabornTable
 
 
 class TimeProfile(object):
